LAMBJ-74 Generator Modularization Refactor
