+++ tests/Compilation/Utils/MSBuildProjectExtensions.cs
@@ -27,7 +27,7 @@ namespace Microsoft.CodeAnalysis.MSBuild
         public static async Task<ImmutableArray<Diagnostic>> GetGeneratorDiagnostics(this Project project)
         {
             var compilation = (await project.GetCompilationAsync())!;
-            var generator = new LambdaGenerator();
+            var generator = new UnitGenerator();
             var driver = CSharpGeneratorDriver.Create(new[] { generator });
             driver.RunGeneratorsAndUpdateCompilation(compilation, out var _, out var diagnotics);
